genai-document-qna/backend/requirements.txt
openai
faiss-cpu
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
numpy
sentence-transformers
torch
